auth-service/requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
pydantic==2.12.5
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.14.0
bcrypt==4.2.0
email-validator==2.2.0
pyjwt==2.9.0
python-jose[cryptography]==3.3.0

